plot api: move common routine to helpers.js
